Segregating responsibilities
