+++ .gitignore
@@ -349,3 +349,4 @@ appsettings.json
 
 # Ignore default JJConsulting logs
 *_applog.txt
+*.DS_Store
