+++ src/JJMasterData.Commons/Dao/Entity/Factory.cs
@@ -50,21 +50,7 @@ public class Factory
         _dataAccess = dataAccess ?? throw new ArgumentNullException(nameof(dataAccess));
     }
 
-    /// <summary>
-    /// Add a record to the database
-    /// </summary>
-    /// <param name="elementName">Dictionary name</param>
-    /// <param name="values">List of values ​​to be stored in the database</param>
-    /// <remarks>
-    /// How to do:
-    /// [key(database field name), value(value to be stored in the database)].
-    /// </remarks>
-    public void Insert(string elementName, Hashtable values)
-    {
-        var element = GetElement(elementName);
-        Insert(element, values);
-    }
-
+    
     /// <summary>
     /// Add a record to the database.
     /// Retorno o id no campo values como referencia
@@ -96,20 +82,6 @@ public class Factory
         }
     }
 
-    /// <summary>
-    /// Update a record in the database .
-    /// </summary>
-    /// <param name="elementName">Dictionary name</param>
-    /// <param name="values">
-    /// List of values ​​to be stored in the database
-    /// [key(database field name), value(value to be stored in the database)]
-    /// </param>
-    /// <returns>Return the number of the rows affected</returns>
-    public int Update(string elementName, Hashtable values)
-    {
-        var element = GetElement(elementName);
-        return Update(element, values);
-    }
 
     /// <summary>
     /// Update a record in the database 
@@ -235,21 +207,6 @@ public class Factory
     /// <summary>
     /// Returns first record based on filter.  
     /// </summary>
-    /// <param name="elementName">Dictionary name</param>
-    /// <param name="filters">List of filters to be used. [key(database field), valor(value stored in database)]</param>
-    /// <returns>
-    /// Return a Hashtable Object. 
-    /// If no record is found then returns null.
-    /// </returns>
-    public Hashtable GetFields(string elementName, Hashtable filters)
-    {
-        var element = GetElement(elementName);
-        return GetFields(element, filters);
-    }
-
-    /// <summary>
-    /// Returns first record based on filter.  
-    /// </summary>
     /// <param name="element">Base element with the basic structure of the table.</param>
     /// <param name="filters">List of filters to be used. [key(database field), valor(value stored in database)]</param>
     /// <returns>
@@ -264,27 +221,7 @@ public class Factory
         return _dataAccess.GetFields(cmd);
     }
 
-    /// <summary>
-    /// Returns records from the database based on the filter.  
-    /// </summary>
-    /// <param name="elementName">Dictionary element name</param>
-    /// <param name="filters">List of filters to be used. [key(database field), valor(value stored in database)]</param>
-    /// <param name="orderby">Record Order, field followed by ASC or DESC</param>
-    /// <param name="regperpage">Number of records to be displayed per page</param>
-    /// <param name="pag">Current page</param>
-    /// <param name="tot">
-    /// If the value is zero, it returns as a reference the number of records based on the filter, 
-    /// running an extra processing in the database, if filled, do not execute the count</param>
-    /// <returns>
-    /// Returns a DataTable with the records found. 
-    /// If no record is found it returns null.
-    /// </returns>
-    public DataTable GetDataTable(string elementName, Hashtable filters, string orderby, int regperpage, int pag,
-        ref int tot)
-    {
-        var element = GetElement(elementName);
-        return GetDataTable(element, filters, orderby, regperpage, pag, ref tot);
-    }
+    
 
     /// <summary>
     /// Returns records from the database based on the filter.    
@@ -342,71 +279,6 @@ public class Factory
         int tot = 0;
         GetDataTable(element, filters, null, 1, 1, ref tot);
         return tot;
-    }
-
-    /// <summary>
-    /// Returns an element with the basic structure of the table
-    /// </summary>
-    /// <param name="elementName">Dictionary name</param>
-    /// <returns></returns>
-    public Element GetElement(string elementName)
-    {
-        if (elementName == null)
-            throw new ArgumentNullException(nameof(elementName), Translate.Key("Invalid dictionary name"));
-
-        var filter = new Hashtable();
-        filter.Add("name", elementName);
-        filter.Add("type", "T");
-
-        var resultElement = GetFields(GetStructure(), filter);
-
-        if (resultElement == null)
-            throw new ArgumentException(Translate.Key("Dictionary {0} not found", elementName));
-
-        var element = JsonConvert.DeserializeObject<Element>(resultElement["json"].ToString());
-        element.Info = resultElement["info"].ToString();
-
-        return element;
-    }
-
-    /// <summary>
-    /// Returns a list of base elements
-    /// </summary>
-    /// <returns></returns>
-    public List<Element> GetListElement()
-    {
-        var filter = new Hashtable();
-        filter.Add("type", "T");
-        return GetListElement(filter, null, 1000, 1);
-    }
-
-    /// <summary>
-    /// Returns a list of base elements
-    /// </summary>
-    /// <param name="filters">List of filters to be used. [key(database field), valor(value stored in database)]</param>
-    /// <param name="orderby">Record Order, field followed by ASC or DESC</param>
-    /// <param name="regporpag">Number of records to be displayed per page</param>
-    /// <param name="pag">Current page</param>
-    /// <returns></returns>
-    public List<Element> GetListElement(Hashtable filters, string orderby, int regporpag, int pag)
-    {
-        var element = GetStructure();
-        if (!ValidateOrderByClause(element, orderby))
-            throw new ArgumentException(Translate.Key("[order by] clause is not valid"));
-
-        var listElement = new List<Element>();
-        int tot = 1000;
-        var dt = GetDataTable(element, filters, orderby, regporpag, pag, ref tot);
-        foreach (DataRow row in dt.Rows)
-        {
-            if (row["type"].ToString().Equals("T"))
-            {
-                Element e = JsonConvert.DeserializeObject<Element>(row["json"].ToString());
-                listElement.Add(e);
-            }
-        }
-
-        return listElement;
     }
 
     /// <summary>
@@ -665,23 +537,5 @@ public class Factory
         return true;
     }
 
-    public Element GetStructure()
-    {
-        var element = new Element(JJService.Settings.TableName, "Data Dictionaries");
-
-        element.Fields.AddPK("type", "Type", FieldType.Varchar, 1, false, FilterMode.Equal);
-        element.Fields["type"].EnableOnDelete = false;
-
-        element.Fields.AddPK("name", "Dictionary Name", FieldType.NVarchar, 64, false, FilterMode.Equal);
-        element.Fields.Add("namefilter", "Dictionary Name", FieldType.NVarchar, 30, false, FilterMode.Contain,
-            FieldBehavior.ViewOnly);
-        element.Fields.Add("tablename", "Table Name", FieldType.NVarchar, 64, false, FilterMode.MultValuesContain);
-        element.Fields.Add("info", "Info", FieldType.NVarchar, 150, false, FilterMode.None);
-        element.Fields.Add("owner", "Owner", FieldType.NVarchar, 64, false, FilterMode.None);
-        element.Fields.Add("sync", "Sync", FieldType.Varchar, 1, false, FilterMode.Equal);
-        element.Fields.Add("modified", "Last Modified", FieldType.DateTime, 15, true, FilterMode.Range);
-        element.Fields.Add("json", "Object", FieldType.Text, 0, false, FilterMode.None);
-
-        return element;
-    }
+  
 }
