+++ src/JJMasterData.Web.Bootstraper/Properties/launchSettings.json
@@ -1,0 +1,28 @@
+{
+  "iisSettings": {
+    "windowsAuthentication": false,
+    "anonymousAuthentication": true,
+    "iisExpress": {
+      "applicationUrl": "http://localhost:9444",
+      "sslPort": 44338
+    }
+  },
+  "profiles": {
+    "JJMasterData.Web.Bootstraper": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "launchUrl": "pt-br/DataDictionary/Element/Index/",
+      "applicationUrl": "https://localhost:7278;http://localhost:5054",
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      }
+    },
+    "IIS Express": {
+      "commandName": "IISExpress",
+      "launchBrowser": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      }
+    }
+  }
+}
